fish app map updated

+++ models/catch.js
@@ -13,11 +13,11 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false
         },
         latitude: {
-            type: DataTypes.DECIMAL(20, 8),
+            type: DataTypes.DOUBLE,
             allowNull: false
         },
         longitude: {
-            type: DataTypes.DECIMAL(20, 8),
+            type: DataTypes.DOUBLE,
             allowNull: false
         }
     });
